HPCC-34153: Move evtool event dump operation to shared library
